REFACTOR-#5916: Align Python engine's API with other engines
