+++ modin/core/execution/python/implementations/pandas_on_python/partitioning/virtual_partition.py
@@ -35,14 +35,6 @@ class PandasOnPythonDataframeAxisPartition(PandasDataframeAxisPartition):
         Whether or not the virtual partition encompasses the whole axis.
     """
 
-    def __init__(self, list_of_blocks, full_axis: bool = True):
-        if not full_axis:
-            raise NotImplementedError(
-                "Pandas on Python execution requires full-axis partitions."
-            )
-
-        self.list_of_block_partitions = list_of_blocks
-
     partition_type = PandasOnPythonDataframePartition
     instance_type = pandas.DataFrame
