+++ modin/pandas/test/test_general.py
@@ -19,7 +19,7 @@ import numpy as np
 from numpy.testing import assert_array_equal
 from modin.utils import get_current_execution, to_pandas
 from modin.test.test_utils import warns_that_defaulting_to_pandas
-from modin.config import Engine, StorageFormat
+from modin.config import StorageFormat
 from pandas.testing import assert_frame_equal
 
 from .utils import (
@@ -661,14 +661,9 @@ def test_unique():
 def test_value_counts(normalize, bins, dropna):
     # We sort indices for Modin and pandas result because of issue #1650
     values = np.array([3, 1, 2, 3, 4, np.nan])
-    with (
-        _nullcontext()
-        if Engine.get() in ["Ray", "Dask", "Unidist"]
-        else warns_that_defaulting_to_pandas()
-    ):
-        modin_result = sort_index_for_equal_values(
-            pd.value_counts(values, normalize=normalize, ascending=False), False
-        )
+    modin_result = sort_index_for_equal_values(
+        pd.value_counts(values, normalize=normalize, ascending=False), False
+    )
     pandas_result = sort_index_for_equal_values(
         pandas.value_counts(values, normalize=normalize, ascending=False), False
     )
@@ -683,14 +678,9 @@ def test_value_counts(normalize, bins, dropna):
     )
     df_equals(modin_result, pandas_result)
 
-    with (
-        _nullcontext()
-        if Engine.get() in ["Ray", "Dask", "Unidist"]
-        else warns_that_defaulting_to_pandas()
-    ):
-        modin_result = sort_index_for_equal_values(
-            pd.value_counts(values, dropna=dropna, ascending=True), True
-        )
+    modin_result = sort_index_for_equal_values(
+        pd.value_counts(values, dropna=dropna, ascending=True), True
+    )
     pandas_result = sort_index_for_equal_values(
         pandas.value_counts(values, dropna=dropna, ascending=True), True
     )
